TMFtrader-src/requirements.txt
shioaji>=1.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=13.0
pandas>=2.2.0
numpy>=1.26.0
loguru>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
yfinance>=0.2.36
